requirements.txt
# FastAPI for API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# LangChain ecosystem
langchain==0.3.7
langchain-core>=0.3.17,<0.4.0
langchain-community==0.3.7
langchain-google-genai==2.0.2
langchain-chroma==0.1.4

# Google GenAI - All we need for embeddings + chat!
google-generativeai>=0.8.0,<0.9.0

# Vector DB - ChromaDB only (lightweight)
chromadb==0.4.22

# File processing (minimal, no OCR)
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
typing-extensions>=4.8.0
pydantic>=2.7.4,<3.0.0
